+++ helpers/ctrlWrapper.js
@@ -1,13 +1,11 @@
 const ctrlWrapper = (ctrl) => {
-  const func = async (req, res, next) => {
+  return async (req, res, next) => {
     try {
       await ctrl(req, res, next);
     } catch (error) {
       next(error);
     }
   };
-
-  return func;
 };
 
 module.exports = ctrlWrapper;
