+++ midelwares/isValidId.js
@@ -5,7 +5,7 @@ const { HttpError } = require('../helpers');
 const isValidId = (req, res, next) => {
   const { id } = req.params;
   if (!isValidObjectId(id)) {
-    next(HttpError(400, `${id} is not valid id`));
+    return next(HttpError(400, `${id} is not valid id`));
   }
   next();
 };
