+++ midelwares/validateBody.js
@@ -4,7 +4,7 @@ const validateBody = (schema) => {
   const func = (req, res, next) => {
     const { error } = schema.validate(req.body);
     if (error) {
-      next(HttpError(400, error.message));
+      return next(HttpError(400, error.message));
     }
     next();
   };
