+++ models/contact.js
@@ -1,4 +1,3 @@
-const Joi = require('joi');
 const { Schema, model } = require('mongoose');
 const { handleMongooseError } = require('../helpers');
 const contactSchema = new Schema(
@@ -22,19 +21,6 @@ const contactSchema = new Schema(
 );
 contactSchema.post('save', handleMongooseError);
 
-const addSchema = Joi.object({
-  name: Joi.string().required(),
-  email: Joi.string().required(),
-  phone: Joi.string().required(),
-});
-const updateFavoriteSchema = Joi.object({
-  favorite: Joi.boolean().required(),
-});
-const schemas = {
-  addSchema,
-  updateFavoriteSchema,
-};
-
 const Contact = model('contact', contactSchema);
 
-module.exports = { Contact, schemas };
+module.exports = { Contact };
