+++ routes/api/contacts.js
@@ -2,18 +2,23 @@ const express = require('express');
 const router = express.Router();
 const ctrl = require('../../controllers/contacts');
 const { validateBody, isValidId } = require('../../midelwares');
-const { schemas } = require('../../models/contact');
+const { schemas } = require('../../schemas/contact');
 
-router.get('/', ctrl.getAll);
-router.get('/:id', isValidId, ctrl.getById);
+router.get('/', ctrl.getAllConcontacts);
+router.get('/:id', isValidId, ctrl.getContactById);
 router.post('/', validateBody(schemas.addSchema), ctrl.contactAdd);
-router.put('/:id', isValidId, validateBody(schemas.addSchema), ctrl.updateById);
+router.put(
+  '/:id',
+  isValidId,
+  validateBody(schemas.addSchema),
+  ctrl.updateContactById
+);
 router.patch(
   '/:id/favorite',
   isValidId,
   validateBody(schemas.updateFavoriteSchema),
   ctrl.updateStatusContact
 );
-router.delete('/:id', isValidId, ctrl.deleteById);
+router.delete('/:id', isValidId, ctrl.deleteContactById);
 
 module.exports = router;
