+++ schemas/contact.js
@@ -1,0 +1,15 @@
+const Joi = require('joi');
+
+const addSchema = Joi.object({
+  name: Joi.string().required(),
+  email: Joi.string().required(),
+  phone: Joi.string().required(),
+});
+const updateFavoriteSchema = Joi.object({
+  favorite: Joi.boolean().required(),
+});
+const schemas = {
+  addSchema,
+  updateFavoriteSchema,
+};
+module.exports = { schemas };
